Finally gave up on goodreads API, Google Books is not great but at least it returns JSON and handles CORS
