+++ app/pages/books/search-books.ts
@@ -1,16 +1,31 @@
 import {Component} from '@angular/core'
-import {GoodReads} from '../../services/goodreads'
+import {GoogleBooks} from '../../services/googlebooks'
+import {Response} from '@angular/http'
+
 
 @Component({
   templateUrl: './build/pages/books/search-books.html',
-  providers: [GoodReads]
+  providers: [GoogleBooks]
 })
 export class SearchBooksPage {
 
+  public results:Array<any>
+
   constructor(
-    public goodreads:GoodReads
+    public googlebooks:GoogleBooks
   ) {
-    this.goodreads.searchBooks('Ancillary')
+    this.googlebooks.searchBooks('Ancillary').subscribe(
+      (res) => {
+        console.log(res)
+        this.results = res['items']
+      },
+      (error:Response) =>  {
+        console.log('Error')
+        console.log(error)
+        console.log(error.status)
+        console.log(error.text())
+      }
+    )
   }
 
 }
